Pflichtfelder eingeführt

+++ Aufgabe3_GhostNetFishing/src/main/java/ghostNetFishing/Netz.java
@@ -46,12 +46,12 @@ public class Netz {
 		this.today = today;
 	}
 
-	public int getID() {
+	public Long getID() {
 		return ID;
 	}
 
 	public void setID(int iD) {
-		ID = iD;
+		ID = (long) iD;
 	}
 
 	public int getLatituteDegree() {
